--- a/PYME/experimental/_triangle_mesh.pxd
+++ b/PYME/experimental/_triangle_mesh.pxd
@@ -111,6 +111,8 @@ cdef class TriangleMesh(TrianglesBase):
     cpdef int edge_flip(self, np.int32_t, bint live_update=*)
     cpdef int edge_collapse(self, np.int32_t, bint live_update=*)
     cpdef int edge_split(self, np.int32_t, bint live_update=*, bint upsample=*)
+    cpdef int relax(self, float l=*, int n=*)
+    cpdef int regularize(self)
     
     #cdef int _insert_new_edge(self, int vertex, int prev=-1, int next=-1, int face=-1, int twin=-1