--- a/PYME/experimental/_triangle_mesh.pxd
+++ b/PYME/experimental/_triangle_mesh.pxd
@@ -95,8 +95,8 @@ cdef class TriangleMesh(TrianglesBase):
     cdef object fix_boundary
     cdef object _manifold
 
-    cdef object _vertex_normals_valid
-    cdef object _face_normals_valid
+    cdef bint _vertex_normals_valid
+    cdef bint _face_normals_valid
 
     cdef object _H
     cdef object _K