Some minor improvements to remesh.

Current stats on `test_er_data.hdf`

```
        1620971 function calls (1606453 primitive calls) in 5.470 seconds

   Ordered by: internal time
   List reduced from 1311 to 20 due to restriction <20>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
    20053    0.847    0.000    0.890    0.000 _triangle_mesh.pyx:910(edge_collapse)
    17561    0.559    0.000    0.952    0.000 _triangle_mesh.pyx:1364(edge_split)
3381/2432    0.477    0.000    0.661    0.000 dual_marching_cubes.py:673(vert_proc)
        5    0.446    0.089    0.499    0.100 _triangle_mesh.pyx:1851(relax)
    15623    0.403    0.000    0.419    0.000 {method 'fix_empty_nodes' of 'PYME.experimental._octree.Octree' objects}
    53447    0.276    0.000    0.276    0.000 {method 'reduce' of 'numpy.ufunc' objects}
       41    0.262    0.006    0.319    0.008 _triangle_mesh.pyx:1793(regularize)
    32208    0.145    0.000    0.422    0.000 dual_marching_cubes.py:140(subdivided)
      976    0.134    0.000    0.134    0.000 weakiddict.py:26(_remover)
    17561    0.126    0.000    0.136    0.000 _triangle_mesh.pyx:1328(_new_vertex)
    40820    0.114    0.000    0.114    0.000 {built-in method numpy.array}
     4986    0.111    0.000    0.527    0.000 dual_marching_cubes.py:175(<listcomp>)
        1    0.100    0.100    0.103    0.103 _triangle_mesh.pyx:524(_initialize_halfedges)
   105366    0.065    0.000    0.098    0.000 _triangle_mesh.pyx:1252(_insert_new_edge)
68879/68252    0.053    0.000    0.517    0.000 {built-in method numpy.core._multiarray_umath.implement_array_function}
    32481    0.052    0.000    0.207    0.000 fromnumeric.py:70(_wrapreduction)
       10    0.051    0.005    0.155    0.016 {built-in method Update}
    19060    0.042    0.000    0.042    0.000 _triangle_mesh.pyx:1517(edge_flip)
       21    0.041    0.002    0.041    0.002 {method 'sort' of 'numpy.ndarray' objects}
   158049    0.037    0.000    0.043    0.000 _triangle_mesh.pyx:1167(_get_insertion_slot)
```

and new stats, with this PR...

```

         1628103 function calls (1613595 primitive calls) in 3.755 seconds

   Ordered by: internal time
   List reduced from 1312 to 20 due to restriction <20>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
3381/2432    0.466    0.000    0.644    0.000 dual_marching_cubes.py:673(vert_proc)
    15623    0.360    0.000    0.375    0.000 {method 'fix_empty_nodes' of 'PYME.experimental._octree.Octree' objects}
       40    0.228    0.006    0.271    0.007 _triangle_mesh.pyx:1904(regularize)
       51    0.203    0.004    0.206    0.004 _triangle_mesh.pyx:445(__get__)
       51    0.202    0.004    0.205    0.004 _triangle_mesh.pyx:426(__get__)
    36021    0.165    0.000    0.165    0.000 {method 'reduce' of 'numpy.ufunc' objects}
    32208    0.142    0.000    0.414    0.000 dual_marching_cubes.py:140(subdivided)
    40810    0.110    0.000    0.110    0.000 {built-in method numpy.array}
    18109    0.109    0.000    0.339    0.000 _triangle_mesh.pyx:1442(edge_split)
      976    0.107    0.000    0.107    0.000 weakiddict.py:26(_remover)
     4986    0.105    0.000    0.479    0.000 dual_marching_cubes.py:175(<listcomp>)
        1    0.101    0.101    0.105    0.105 _triangle_mesh.pyx:524(_initialize_halfedges)
    20670    0.101    0.000    0.140    0.000 _triangle_mesh.pyx:910(edge_collapse)
    18109    0.096    0.000    0.104    0.000 _triangle_mesh.pyx:1406(_new_vertex)
   108654    0.059    0.000    0.088    0.000 _triangle_mesh.pyx:1330(_insert_new_edge)
        5    0.055    0.011    0.487    0.097 _triangle_mesh.pyx:1962(relax)
    32616    0.052    0.000    0.206    0.000 fromnumeric.py:70(_wrapreduction)
68974/68357    0.048    0.000    0.490    0.000 {built-in method numpy.core._multiarray_umath.implement_array_function}
       10    0.038    0.004    0.118    0.012 {built-in method Update}
   162981    0.031    0.000    0.038    0.000 _triangle_mesh.pyx:1245(_get_insertion_slot)
```