@zacsimile 2021-11-30
Actually I just made things a lot faster by switching 

```
score_pre = abs(v1) + abs(v2) + abs(v3) + abs(v4)
score_post = abs(v1-1) + abs(v2-1) + abs(v3+1) + abs(v4+1)
```

to 

```
score_pre = v1*v1 + v2*v2 + v3*v3 + v4*v4
score_post = (v1-1)*(v1-1) + (v2-1)*(v2-1) + (v3+1)*(v3+1) + (v4+1)*(v4+1)
```

Since we only check `score_post < score_pre`, I think this should be OK.