@zacsimile 2021-11-30
New benchmark with the regularization fix on `test_er_data_uncropped.hdf`.

Master branch:

```
         19697520 function calls (19651348 primitive calls) in 80.709 seconds

   Ordered by: internal time
   List reduced from 1311 to 20 due to restriction <20>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
   397262   17.228    0.000   18.104    0.000 _triangle_mesh.pyx:910(edge_collapse)
        5   12.130    2.426   13.300    2.660 _triangle_mesh.pyx:1851(relax)
   350017   10.430    0.000   18.167    0.000 _triangle_mesh.pyx:1364(edge_split)
       47   10.397    0.221   11.791    0.251 _triangle_mesh.pyx:1793(regularize)
   529547    2.971    0.000    2.971    0.000 {method 'reduce' of 'numpy.ufunc' objects}
17956/12627    2.822    0.000    3.932    0.000 dual_marching_cubes.py:673(vert_proc)
   350017    2.461    0.000    2.775    0.000 _triangle_mesh.pyx:1328(_new_vertex)
        1    2.088    2.088    2.142    2.142 _triangle_mesh.pyx:524(_initialize_halfedges)
    84961    2.037    0.000    2.128    0.000 {method 'fix_empty_nodes' of 'PYME.experimental._octree.Octree' objects}
   377478    1.370    0.000    1.370    0.000 _triangle_mesh.pyx:1517(edge_flip)
  2100102    1.286    0.000    1.955    0.000 _triangle_mesh.pyx:1252(_insert_new_edge)
    26860    0.912    0.000    2.992    0.000 dual_marching_cubes.py:175(<listcomp>)
   219196    0.869    0.000    0.869    0.000 {built-in method numpy.array}
  3150153    0.851    0.000    0.959    0.000 _triangle_mesh.pyx:1167(_get_insertion_slot)
      237    0.795    0.003    0.795    0.003 {method 'sort' of 'numpy.ndarray' objects}
   161160    0.759    0.000    2.185    0.000 dual_marching_cubes.py:140(subdivided)
       12    0.682    0.057    0.682    0.057 {method 'argsort' of 'numpy.ndarray' objects}
        6    0.660    0.110    0.664    0.111 _triangle_mesh.pyx:445(__get__)
        5    0.659    0.132   18.764    3.753 _triangle_mesh.pyx:2012(collapse_edges)
        6    0.518    0.086    0.519    0.087 _triangle_mesh.pyx:426(__get__)
```

This PR:
```
         19766522 function calls (19720360 primitive calls) in 48.857 seconds

   Ordered by: internal time
   List reduced from 1312 to 20 due to restriction <20>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
       51    6.270    0.123    6.297    0.123 _triangle_mesh.pyx:445(__get__)
       51    4.994    0.098    5.006    0.098 _triangle_mesh.pyx:426(__get__)
       49    4.172    0.085    6.196    0.126 _triangle_mesh.pyx:1904(regularize)
   399026    2.999    0.000    3.960    0.000 _triangle_mesh.pyx:910(edge_collapse)
   354090    2.755    0.000    7.397    0.000 _triangle_mesh.pyx:1442(edge_split)
17956/12627    2.607    0.000    3.614    0.000 dual_marching_cubes.py:673(vert_proc)
        5    2.047    0.409   13.721    2.744 _triangle_mesh.pyx:1965(relax)
        1    2.027    2.027    2.079    2.079 _triangle_mesh.pyx:524(_initialize_halfedges)
   633022    1.995    0.000    1.995    0.000 _triangle_mesh.pyx:1628(edge_flip)
   354090    1.937    0.000    2.091    0.000 _triangle_mesh.pyx:1406(_new_vertex)
    84961    1.872    0.000    1.957    0.000 {method 'fix_empty_nodes' of 'PYME.experimental._octree.Octree' objects}
  2124540    1.204    0.000    1.794    0.000 _triangle_mesh.pyx:1330(_insert_new_edge)
   179665    0.844    0.000    0.844    0.000 {method 'reduce' of 'numpy.ufunc' objects}
    26860    0.821    0.000    2.729    0.000 dual_marching_cubes.py:175(<listcomp>)
   219186    0.813    0.000    0.813    0.000 {built-in method numpy.array}
      237    0.755    0.003    0.755    0.003 {method 'sort' of 'numpy.ndarray' objects}
       12    0.691    0.058    0.691    0.058 {method 'argsort' of 'numpy.ndarray' objects}
   161160    0.690    0.000    1.942    0.000 dual_marching_cubes.py:140(subdivided)
  3186810    0.629    0.000    0.736    0.000 _triangle_mesh.pyx:1245(_get_insertion_slot)
       51    0.487    0.010    6.791    0.133 _triangle_mesh.pyx:455(__get__)
```