@David-Baddeley 2021-11-30
I don't think you can make that abs -> square substitution and keep the maths the same - it would be fine if the `abs()` was around the outside of the sum, but because it's a sum of abs()s I don't think it will be OK.

suspect the issue here is that it's calling the python abs function, rather than c `abs()` from stdlib (which the compiler should make short work of)