@David-Baddeley 2021-11-30
You should be able to see in the generated code which `abs` it ends up calling - `PyNumber_Absolute` or something similar is what you need to look for (bad)