@David-Baddeley 2021-11-30
if it's calling a proper abs, then the speed improvement is likely due to avoiding splits/collapses that ti would normally do rather than any actual speed improvement.