@David-Baddeley 2021-11-30
looking at an old version, it definitely used `PyNumber_Absolute` ...

![image](https://user-images.githubusercontent.com/19475296/144138885-978e3d4f-99ee-4894-bc13-c6a515dcc65c.png)