@zacsimile 2021-11-30
Yep, |x| -> x^2 won't work since it's a sum of x^2 terms. I got too excited.

I'd love to switch to C version of `abs()`. Do you know how to force cython to do this during compilation? I figured it would do so automatically, but it seems not.