@David-Baddeley 2021-11-30
changing the type definition to `int` from `np.int_32t` at the top of the function fixes it cython uses the stdlib abs instead ...