@David-Baddeley 2021-11-30
![image](https://user-images.githubusercontent.com/19475296/144140448-1e194cef-d8ba-47d3-9b9e-4eef7ebf04cd.png)

now if only we could get rid of that `if(unlikely(....)` stuff