@zacsimile 2021-11-30
Switched to `int` to force c abs, still a nice improvement.