@zacsimile 2021-11-30
How did you visualize the C calls?