@David-Baddeley 2021-11-30
the .html output is automatically generated when you run `cython` (tecnically it's the -a command line option - i.e. `cython -a myfile.pyx`), but for some reason I don't need to specify the -a on my machine.