@David-Baddeley 2021-11-30
It's super useful, as it highlights any line which results in a python call.