@David-Baddeley 2021-11-30
Looking at the profile, my next targets would probably be `face_normals` and `vertex_neighbours`. 

There looks to be a lot we could do here - low hanging fruit would be to just keep a boolean flag for, e.g. `_face_normals_valid` and invalidate this rather than always doing the `np.all(...)` check, but we could probably also look at making `c_update_face_normals` accept something other than a python list.