@zacsimile 2021-11-30
Yeah those are both good ideas. The boolean certainly could be implemented quite quickly and would help a bit. Maybe a `c_update_all_face_normals()` function instead of eliminating the ability to past a list of indices?