@zacsimile 2021-11-30
I added the flags. I didn't see any noticeable improvements but maybe eliminating the list will help.

I actually don't see `face_normals` or `vertex_neighbors` showing up high in the profile. What are you sorting by?