@David-Baddeley 2021-11-30
They're the top two lines ....

```
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
       51    6.079    0.119    6.107    0.120 _triangle_mesh.pyx:445(__get__)
       51    4.974    0.098    4.985    0.098 _triangle_mesh.pyx:426(__get__)
```